+++ Assets/Scripts/Debug/TextField.cs
@@ -6,11 +6,13 @@ public class TextField : MonoBehaviour
 {
     //private PlayerMaster pMaster = new PlayerMaster();
     private EnemyMaster eMaster = new EnemyMaster();
+
+    private ShotUnitMaster shotMaster = new ShotUnitMaster();
     void OnGUI()
     {
         // テキストフィールドを表示する
         GUI.Label(new Rect(10, 10, 200, 200),"DebuData","box");
-        GUI.Label(new Rect(13, 25, 200, 20), "Position:" + eMaster.Destinaion.ToString());
+        GUI.Label(new Rect(13, 25, 200, 20), "Rotation:" + shotMaster.Rotation.ToString());
         //GUI.Label(new Rect(13, 40, 200, 20), "Speed   :" + pMaster.Speed.ToString());*/
 
     }
