+++ Assets/Scripts/Player/ShotUnitMaster.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class ShotUnitMaster : MonoBehaviour
+{
+    private static Quaternion rotation;
+
+    public Quaternion Rotation { get => rotation; set => rotation = value; }
+}
