+++ Assets/Scripts/Player/ShotUnitMaster.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f31047da6311c574dbc4c127fbfbbe45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
