+++ Assets/Scripts/SkyFile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 194605f055c22164c861fa0c329f97ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
