+++ Assets/Scripts/SkyFile/ClearControl.cs
