+++ Assets/Scripts/SkyFile/PlayerHpBar.cs.meta
